- dhis logic sub classes - Pin dialog crash fix - android sdk update
