+++ app/src/main/assets/paperwork.json
@@ -1,1 +1,1 @@
-{"buildTime":"2020-07-09 13:44","gitSha":"eb5a8c7ff"}
+{"buildTime":"2020-07-24 13:11","gitSha":"a8d8cd71c"}
